frontend/requirements.txt
flask==2.3.3
werkzeug==2.3.7
python-dotenv==1.0.0
PyPDF2==3.0.1
requests==2.31.0
crewai==0.28.5 